strlen prototypes

+++ ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: cschuijt <cschuijt@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/10 14:28:09 by cschuijt      #+#    #+#                 */
-/*   Updated: 2022/10/12 16:41:19 by cschuijt      ########   odam.nl         */
+/*   Updated: 2022/10/12 17:21:29 by cschuijt      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int		ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
