+++ ft_strnstr.c
@@ -6,14 +6,14 @@
 /*   By: cschuijt <cschuijt@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/11 11:59:46 by cschuijt      #+#    #+#                 */
-/*   Updated: 2022/10/11 13:17:06 by cschuijt      ########   odam.nl         */
+/*   Updated: 2022/10/12 17:22:11 by cschuijt      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n);
-int	ft_strlen(const char *s);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
+size_t	ft_strlen(const char *s);
 
 char	*ft_strnstr(const char *hay, const char *needle, size_t len)
 {
